JAMES-2897 Avoid an interface joining together the two mailboxDAO
The older one is not supposed to be written to.
